category-prediction/requirements.txt
numpy
gunicorn
Flask
tabulate
seaborn
scikit-learn
wordcloud
deep_translator
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
enum-compat==0.0.3
filelock==3.13.1
fsspec==2023.12.2
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.16.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.1
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
sympy==1.12
tokenizers==0.15.1
torch==2.2.0
torch-model-archiver==0.7.1
torchaudio==2.2.0
torchserve==0.9.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.37.0
typing_extensions==4.9.0
urllib3==2.1.0
wheel==0.42.0
